+++ package.json
@@ -333,8 +333,8 @@
           "type": "tree"
         },
         {
-          "id": "runs",
-          "name": "Runs",
+          "id": "history",
+          "name": "Run History",
           "type": "tree"
         }
       ]
